Добавил базовый класс для элементов справочника. От него наследуются Language, Term, Trend и Profession
