--- .idea/misc.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="DesignSurface">
-    <option name="filePathToZoomLevelMap">
-      <map>
-        <entry key="..\:/Projects/IT_English/app/src/main/res/drawable/ic_action_refresh.xml" value="0.11302083333333333" />
-        <entry key="..\:/Projects/IT_English/app/src/main/res/layout/activity_language.xml" value="0.1539855072463768" />
-        <entry key="..\:/Projects/IT_English/app/src/main/res/layout/activity_main.xml" value="0.1" />
-        <entry key="..\:/Projects/IT_English/app/src/main/res/layout/activity_term.xml" value="0.16802536231884058" />
-        <entry key="..\:/Projects/IT_English/app/src/main/res/layout/fragment_terms.xml" value="0.1" />
-        <entry key="..\:/Projects/IT_English/app/src/main/res/layout/fragment_trends.xml" value="0.1" />
-        <entry key="..\:/Projects/IT_English/app/src/main/res/layout/template.xml" value="0.1" />
-        <entry key="..\:/Projects/IT_English/app/src/main/res/menu/app_bar.xml" value="0.18" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/drawable-v24/rounded_layout.xml" value="0.26153846153846155" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/drawable/customborder.xml" value="0.1" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/drawable/customfon.xml" value="0.23425925925925925" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/drawable/ic_launcher_background.xml" value="0.1" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/layout/activity_language.xml" value="0.2" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/layout/activity_main.xml" value="0.264" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/layout/activity_profession.xml" value="0.264" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/layout/activity_term.xml" value="0.39473684210526316" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/layout/activity_trend.xml" value="0.4" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/layout/fragment_languages.xml" value="0.2259963768115942" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/layout/fragment_notifications.xml" value="0.2259963768115942" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/layout/fragment_profession.xml" value="0.2259963768115942" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/layout/fragment_terms.xml" value="0.2259963768115942" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/layout/fragment_trends.xml" value="0.2259963768115942" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/layout/template.xml" value="0.536" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/layout/term.xml" value="0.6" />
-        <entry key="..\:/Users/Asus/Documents/GitHub/IT_English/app/src/main/res/menu/bottom_nav_menu.xml" value="0.25" />
-      </map>
-    </option>
-  </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" project-jdk-name="1.8" project-jdk-type="JavaSDK">
-    <output url="file://$PROJECT_DIR$/build/classes" />
-  </component>
-  <component name="ProjectType">
-    <option name="id" value="Android" />
-  </component>
-</project>
